Moves types to global
